Improve _download_img() method for SectionDataSet
Previously this method was using mkstemp to create both
a file object handle as well as a filename then immediately
closing the file object. A more straightforward strategy is
to use a NamedTemporaryFile and 'with' context to
accomplish the same thing.

Also add more type annotations and regenerate the relevant notebook.